backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn[standard]
browser-use #is installed manually in Dockerfile, REMOVE any line for it here
playwright
ollama
python-dotenv
json-repair
langchain-ollama
# pyperclip==1.9.0 # Remove if not used